requirements.txt
torch
torchvision
scikit-learn
tqdm
wandb
python-dotenv
pillow
transformers
streamlit
mlflow
opencv-python
pandas
numpy
langchain
langchain-core
langchain-openai
langchain-google-genai
langchain-google-vertexai 